+++ src/reflex_django/state/views/meta.py
@@ -46,7 +46,6 @@ class ModelCRUDMeta:
     backend_class: ClassVar[type | None] = None
     structured_errors: ClassVar[bool] = False
     run_model_validation: ClassVar[bool] = False  # enables validate_model_full_clean()
-    load_context_processors: ClassVar[bool] = True
     reset_after_save: ClassVar[bool] = True
     form_reset_var: ClassVar[str | None] = "form_reset_key"
     use_form_submit: ClassVar[bool] = False
@@ -78,7 +77,6 @@ class ModelListMeta:
     login_required_actions: ClassVar[frozenset[str] | None] = None
     permission_classes: ClassVar[tuple[type, ...]] = ()
     backend_class: ClassVar[type | None] = None
-    load_context_processors: ClassVar[bool] = True
     paginate_by: ClassVar[int | None] = None
     max_page_size: ClassVar[int] = 100
     page_var: ClassVar[str] = "page"